Started adding some messy tests. Will need to clean up.
At the moment, the test include loading data into the db, which is great, and needs to be tested, but it might also be a good idea to be run tests on a chado instance that has been loaded with the yeast test data. This would ensure that the bindings work fully with the GMOD schema.